# api/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hanasand Public Intelligence API",
    "version": "1.0.0",
    "description": "Versioned, metadata-safe access to production-backed public intelligence. Version 1 is read-only: POST search operations are retrieval operations, and no mutation or idempotency-key contract is advertised.",
    "x-compatibility-policy": "Additive fields and endpoints may be introduced within v1. Removing or changing existing fields, semantics, authentication, or pagination requires a new major API version.",
    "x-idempotency-policy": "v1 exposes no public mutations. Search POSTs are retrieval-only and safe to retry; no Idempotency-Key header is required."
  },
  "servers": [
    {
      "url": "https://api.hanasand.com/api/v1",
      "description": "Production"
    }
  ],
...
  "paths": {
    "/ti/search": {
      "post": {
        "operationId": "searchThreatIntelligence",
        "summary": "Search public threat intelligence",
        "description": "One bounded search is available anonymously. Results are never shared-cacheable.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search result",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store, max-age=0"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/ti/search/batch": {
      "post": {
        "operationId": "batchSearchThreatIntelligence",
        "summary": "Search up to 25 unique queries",
        "description": "Requires an API key or authenticated session. HTTP 207 reports per-query partial failures without disguising them as successful search results.",
        "security": [
          {
            "ApiKey": []
          },
          {
            "SessionBearer": [],
            "SessionId": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "All searches completed",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store, max-age=0"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSearchResponse"
                }
              }
            }
          },
          "207": {
            "description": "One or more searches failed",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store, max-age=0"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchSearchResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
    "/actors": {
      "get": {
        "operationId": "listActors",
        "summary": "Actor intelligence profiles",
        "security": [
          {
            "ApiKey": []
          },
          {
            "SessionBearer": [],
            "SessionId": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Case-insensitive text filter.",
            "schema": {
              "type": "string",
              "maxLength": 200
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Records per page.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Opaque cursor returned by the preceding page.",
            "schema": {
              "type": "string",
              "pattern": "^\\d+$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Actor page",
            "headers": {
              "X-Request-Id": {
                "schema": {
                  "type": "string"
                }
              },
              "Cache-Control": {
                "schema": {
                  "type": "string",
                  "const": "no-store, max-age=0"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActorCollection"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/RateLimited"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Provisioned key with an exact method and route scope."
      },
      "SessionBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "session-token"
      },
      "SessionId": {
        "type": "apiKey",
        "in": "header",
        "name": "id",
        "description": "Session identifier required together with SessionBearer."
      }
    },
    "schemas": {
      "ErrorDetail": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ]
      },
      "Pagination": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "nextCursor": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "limit",
          "total",
          "nextCursor"
        ]
      },
      "Meta": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ]
      },
      "SearchRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "query": {
            "type": "string",
            "minLength": 2,
            "maxLength": 200,
            "example": "APT29"
          }
        },
        "required": [
          "query"
        ]
      },
      "BatchSearchRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "queries": {
            "type": "array",
            "minItems": 1,
            "maxItems": 25,
            "items": {
              "type": "string",
              "minLength": 2,
              "maxLength": 200
            },
            "example": [
              "APT29",
              "CVE-2024-3094"
            ]
          }
        },
        "required": [
          "queries"
        ]
      },
      "Activity": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "date": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "sourceIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "claimType": {
            "type": "string"
          },
          "victimName": {
            "type": "string"
          },
          "affectedSectors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "countries": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "impact": {
            "type": "string"
          },
          "firstReportedAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastReportedAt": {
            "type": "string",
            "format": "date-time"
          },
          "publisherCount": {
            "type": "integer",
            "minimum": 0
          },
          "corroboratingSourceIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "contradictingSourceIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "date",
          "title",
          "detail",
          "confidence",
          "sourceIds"
        ]
      },
      "Target": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "sector": {
            "type": "string"
          },
          "regions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "rationale": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        },
        "required": [
          "sector",
          "regions",
          "rationale",
          "confidence"
        ]
      },
      "Ttp": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "attackId": {
            "type": "string"
          },
          "tactic": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        },
        "required": [
          "name",
          "tactic",
          "detail",
          "confidence"
        ]
      },
      "Dataset": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "clear_web",
              "public_channel",
              "darknet_metadata",
              "vendor_report",
              "stix_export"
            ]
          },
          "coverage": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "available",
              "planned",
              "metadata_only"
            ]
          },
          "url": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "name",
          "type",
          "coverage",
          "status"
        ]
      },
      "SearchSource": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "provenance": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri"
          },
          "captureId": {
            "type": "string"
          },
          "sourceRequestId": {
            "type": "string"
          },
          "sourceFamily": {
            "type": "string"
          },
          "parserStatus": {
            "type": "string"
          },
          "reportDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastCollectedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "provenance"
        ]
      },
      "ActorIntelligence": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "actorClass": {
            "type": "string"
          },
          "attribution": {
            "type": "string"
          },
          "firstSeen": {
            "type": "string",
            "format": "date-time"
          },
          "lastSeen": {
            "type": "string",
            "format": "date-time"
          },
          "motivation": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "malwareTools": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "campaigns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "infrastructure": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetSectors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "geographies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "confidenceReasoning": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "sourceProvenance": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "WatchlistCandidate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "company",
              "domain",
              "vendor"
            ]
          },
          "value": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          }
        },
        "required": [
          "kind",
          "value",
          "reason"
        ]
      },
      "Actionability": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schemaVersion": {
            "type": "string",
            "const": "ti.query.actionability.v1"
          },
          "alertDisposition": {
            "type": "string",
            "enum": [
              "ready_for_alert_review",
              "watchlist_required",
              "case_ready",
              "not_alertable",
              "needs_enrichment"
            ]
          },
          "shouldAlert": {
            "type": "boolean"
          },
          "rationale": {
            "type": "string"
          },
          "watchlistCandidates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WatchlistCandidate"
            }
          }
        }
      },
      "SearchResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "query": {
            "type": "string"
          },
          "queryKind": {
            "type": "string",
            "enum": [
              "actor",
              "domain",
              "cve",
              "indicator",
              "organization",
              "free_text"
            ]
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "mode": {
            "type": "string",
            "enum": [
              "scraper",
              "seeded",
              "live_search"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "searching",
              "partial",
              "ready",
              "metadata_review",
              "blocked_unsafe_target",
              "needs_source_activation"
            ]
          },
          "runId": {
            "type": "string"
          },
          "refreshAfterSeconds": {
            "type": "integer",
            "minimum": 0
          },
          "summary": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "lastSeen": {
            "type": "string"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recentActivity": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Activity"
            }
          },
          "targets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Target"
            }
          },
          "ttps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Ttp"
            }
          },
          "datasets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Dataset"
            }
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchSource"
            }
          },
          "notes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "actorIntelligence": {
            "$ref": "#/components/schemas/ActorIntelligence"
          },
          "actionability": {
            "$ref": "#/components/schemas/Actionability"
          }
        },
        "required": [
          "query",
          "generatedAt",
          "mode",
          "summary",
          "confidence",
          "lastSeen",
          "aliases",
          "recentActivity",
          "targets",
          "ttps",
          "datasets",
          "sources",
          "notes"
        ]
      },
      "BatchSuccess": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "query": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "const": "ok"
          },
          "result": {
            "$ref": "#/components/schemas/SearchResult"
          }
        },
        "required": [
          "query",
          "status",
          "result"
        ]
      },
      "BatchFailure": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "query": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "const": "error"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "query",
          "status",
          "error"
        ]
      },
      "BatchSearchResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "count": {
            "type": "integer",
            "minimum": 1,
            "maximum": 25
          },
          "partial": {
            "type": "boolean"
          },
          "results": {
            "type": "array",
            "items": {
              "oneOf": [
                {
                  "$ref": "#/components/schemas/BatchSuccess"
                },
                {
                  "$ref": "#/components/schemas/BatchFailure"
                }
              ]
            }
          }
        },
        "required": [
          "generatedAt",
          "count",
          "partial",
          "results"
        ]
      },
      "Actor": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "canonicalName": {
            "type": "string"
          },
          "normalizedName": {
            "type": "string"
          },
          "actorType": {
            "type": "string"
          },
          "aliases": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "firstSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          },
          "evidenceCount": {
            "type": "integer",
            "minimum": 0
          },
          "sourceIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "captureIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reviewState": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "canonicalName",
          "aliases",
          "confidence",
          "sourceIds",
          "captureIds"
        ]
      },
...
      "ActorCollection": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Actor"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "pagination",
          "meta"
        ]
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid input",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store, max-age=0"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store, max-age=0"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Credential scope does not permit this operation",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store, max-age=0"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "RateLimited": {
        "description": "Rate limit exceeded",
        "headers": {
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          "X-Rate-Limit-Remaining": {
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "InternalError": {
        "description": "Unexpected request processing failure",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store, max-age=0"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Upstream intelligence service unavailable",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string"
            }
          },
          "Cache-Control": {
            "schema": {
              "type": "string",
              "const": "no-store, max-age=0"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}